cmd/helper: move metric logic inside its own package
